Initial commit of source files to prevent data loss due to system failures. At this point, these files allow you to act as a C-ECHO SCU.
The Dicom/Network project wraps the lower-level SWIG wrappers and constitutes as the API to the DICOM networking functionality in the Framework.

The Common/OFFIS/dcmtk-3.5.3/dcmnet/SWIG_Wrapper project contains the SWIG interface file that generates the wrapper files. The dcmnet_wrap.cxx file is compiled right in the dcmnet/libsrc project, using the dcmnet_wrapper.csproj project file. The result of this compile is then used in the creation of the C# wrapper types, under the csharp folder.

The Test project contains a test scaffold to exercise the simple cases of the functionality exposed by the ClearCanvas.Dicom.Network namespace.

Some reorganization of the directory structures is probably in order. This is an interim commit that will allow my work to be backed up and saved. The source code, however, does build successfully, and works.