Reorganizations of the source structure in order to combine the Data and Network assemblies into one assembly.
The code now builds, but it hasn't been tested to see if it still works.

Also created a new DicomException class in the ClearCanvas.Dicom.Exceptions namespace that can be thrown by any entity that includes the ClearCanvas.Dicom.Exceptions namespace will be able to catch or throw this the DicomException class.

Otherwise, undesirable situations like the following could arise: The Network OFFIS wrapper classes must throw the DicomException class. The client code would like to catch such an exception. This means that the client code will have to use the C# abstraction classes, as well as include the lower-level OFFIS wrappers just to catch this exception.