Added in SWIG %typemaps that will store a C# reference of certain objects passed in, using a member variable in the specific C# proxy class.
This is necessary per 17.4.2 of SWIG manual, since the underlying C/C++ code may still depend on the object's existence, but the GC may collect the object, because it is not aware of any outstanding references to the object.

See T_ASC_Network, with the creation of an association using association parameters. From the C# point of view, once the association parameters have been used, it is no longer referenced (in the C# world -- see DicomClient.Retrieve and DicomClient.Query). In reality, the association parameters become a member of the association object, and therefore, must remain valid for all subsequent association commands.

Note that for the %typemaps to work in cooperation with the %extend (that provides the CreateAssociation method), the %typemaps have to appear before the declaration of the struct/class that %extend applies to.