Requirement: R197
Altered the action model code so that the ActionModelRoot node corresponds to the Site of the actions in the model, rather than having the site as a child of the root.  This formalizes a 1 to 1 correspondence between an action model and a site, which should help to simplify the understanding and usage of action models.