Requirement: R787
Extended code-generator to obey the "access" attribute in the hibernate mapping file.  Thus, if access="property" (or omitted), both a getter and setter will be generated.  If access="nosetter.namingstrategy", only a getter will be generated.  If access="field.namingstrategy", only the field will be generated - no property will be generated at all.

Note that the code generator ignores the naming strategy portion of the attribute, since we always use camelcase-underscore, this is the only strategy that is supported.