Bug: #77 - Unexpected state change when drawing measurement tools
This had actually been fixed when mouse capture was put in initially.  However, now that each tile is its own UserControl, mouse capture had to be restored because it no longer worked with the new design.  Functionally, it is basically the same as before, but with some new enhancements based on discussions with Wes, Norman & Dave:[[BR]]

- The Tile (UserControl) keeps mouse capture unconditionally until the graphic is completed by the user.  Even ROI graphics using atomic mouse clicks will keep the capture until the graphic itself releases it.
- ROI graphics are not restricted to stay within the tile.  The user can drop anchors wherever he/she wishes (except outside the desktop window boundaries).  If a developer wants to change this, he/she can override the appropriate methods in the state objects to restrict the control points.
- As a result of these changes, you cannot click on any buttons (toolbar, etc) until the graphic is finished.  Again, developers can override the appropriate members to change this behaviour and release the capture.