Requirement: R1367
Modified WorklistComponent so that instead of having the view listen to the SelectionChanged event on the TableView and call WorklistComponent.SetSelection(...), the TableView.Selection property is now simply bound to a corresponding WorklistComponent.Selection property.  This provides a slightly simpler paradigm for interfacing with the TableView.  In theory it should also allow the TableView selection to be set programatically, although this has not been tested.

I'm not going to bother retrofitting any other existing components to use binding, so its basically up to whoever "owns" those components to do so if they wish.  Otherwise, the old way of doing things still works, and you can use the new way going forward.