Bug: #228
Part 2: Made the mapping between ButtonAction/MenuAction and the actual ToolStripItems flexible by introducing the ButtonActionView and MenuActionView classes, and using the AssociateView attribute.  This means that arbitrary controls can now be added to the toolbar, simply by defining a new subclass of Action and an associated view that returns a ToolStripItem as its GuiElement.

This completes the feature request.