Bug: #228
The previous commit for this bug fix introduced a new bug, because the ClickAction class did not have an associated view.  Changed ToolStripBuilder so that any implementation of IClickAction that does not have an associated view attribute will have the appropriate default menu/toolbar toolstrip items constructed directly.  This also serves as an optimization: since 99% of actions will probably be simple click actions, it is faster to construct these items directly rather than through the associate view mechanism.  However, if an AssociateView attribute is present, it will take precendence, thereby allowing for the flexibility of extending the ClickAction class and supplying a custom view.