Requirement: R1427
Modified StandardSettingsProvider so that it can resolve default values from embedded resources.  The strategy by which is does this is very straightforward:

If the default value supplied by the DefaultSettingValueAttribute looks like it could be the name of a resource (e.g has the form *.*.* etc), then the provider looks for an embedded resource by that name.  If it finds one, it returns the contents of the embedded resource as the default value.  If it does not find one, then it returns the raw default value.