Requirement: R1811 Interim commit. ShredHost can now run inside of a Windows service and be controlled from the Service Applet. A limitation, however, is that Shreds HAVE to reference the strong name signed version of ClearCanvas.Common.dll. This is not so good, because this basically means that any other ClearCanvas Framework assemblies the Shred references, must also be strong name signed, since its likely that it too will reference ClearCanvas.Common.dll.
The reason that the Shred has to reference the SNS version of ClearCanvas.Common.dll, is to allow it to be recognized as a plugin. Basically, ShredHost uses the SNS version of ClearCanvas.Common.dll to scan for plugins, because it has to. If the shred was built against a private version of ClearCanvas.Common.dll, the PluginAttribute types from the two different assemblies won't match, and the shred won't be recognized as a plugin.

If this sounds confusing, ask me for a one-on-one explanation, because I'm too tired to write any more.

Further notes: An msbuild script has been written to create SNS versions of ShredHostService, ShredHost and Common. This needs to be done, since we don't have want to modify the existing csproj files, in order to delay-sign these assemblies.