Requirement: R1907
Implemented basic auditing.  Tried to use the pattern found in "Hibernate in Action", which suggests auditing on a separate Session that enlists in the same transaction as the primary session.  However, NHibernate does not seem to like this approach (the book is based on the java version).  So for now, auditing occurs on the same Session as the rest of the operation - the limitation of this is that Read Operations cannot be audited.

The contents of the audit record can be controlled by providing a custom implementation of ITransactionRecorder.  For the time being however, a class called DefaultTransactionRecorder is hard-coded into the AuditAdvice class.  In future, the ReadOperationAttribute and UpdateOperationAttribute classes could have a parameter that specifies a custom implementation of ITransactionRecorder, which would be picked up by the AuditAdvice, and which would allow a given operation to be audited in a customized way.