Requirement: R1861
Added CodedValue, CodedValueDictionary, and CodedValueHbm.  Together these classes provide support for coded-values, which are similar to enumerations.  The main difference is that the coded-value inherently supports the notion of a mapping between a code and a value, and coded-value sets are configurable rather than hard-coded (although it is still possible to hard-code values if desired).

More documentation will follow at some point.