Bug: #315
Currently, the RIS and ImageViewer both have PostBuild.bat files in the same directory as their solution files.
If you want to have more than one solution file in the directory (for different solution configurations), this is a problem, since there will be a name clash.  To fix this, I've changed the calling PostBuild.bat (located in Executable) so that it calls a batch file with the same name of the solution.  For example, in the ImageViewer directory, I can now have something like this:

ImageViewer.sln
ImageViewer.bat

ImageViewerVolume.sln
ImageViewerVolume.bat

The ImageViewer.sln is the standard configuration of the viewer and ImageViewer.bat copies over the standard files.

ImageViewerVolume.sln has the additional volume projects and ImageViewerVolume.bat copies over the standard files plus the additional volume dlls.

I've made the appropriate change in /RIS by renaming the PostBuild.bat in that directory to Ris.bat.