Bug: #157
Refactored this component to have the concept of a LocalDataStoreNode in the ServerTree. The main work done here is to remove several utility classes. Now, the classes that are used to manage the ServerTree are directly used to persist the ServerTree to an XML file. They are also directly used by the Navigator component as the model behind the visual tree view UI. Everything to do with ServerTree is contained in ServerTree.cs.

The UI code has not been re-architected, except to make the new classes work, as that would have been outside the scope of this bug. Some notes for users of the ServerTree:

1. Instead of accessing the DicomAE property of a node, to get the ClearCanvas.Dicom.Network.ApplicationEntity subobject, use GetApplicationEntity() instead.
2. Each ServerGroup node now contains two different Lists<>, one for child groups and one for child servers. This separation makes the code easier to understand and manage.
3. The Root node contains two nodes: LocalDataStoreNode and a ServerGroupNode. All new servers and server groups will be added to this ServerGroupNode.
4. Logic for checking whether or not a group or server can be drag+dropped on another, has been encapsulated in the nodes themselves, in the CanAddAsChild() call.

Unit tests have been disabled from the compile for the time being.