Requirement: R2437
Completed HQL report query generator.  This is an extension to the original ruby codegen, and allows brokers, criteria, and result classes to be generated in C# based upon an XML model of the query.  The XML is stored in a file with the extension *.hrq.xml, which is similar to *.hbm.xml (where hrq stands for Hql Report Query).

Outstanding:
1. Although the generator works, it does not provide good diagnostics (imagine a compiler with meaningless error messages), which means that errors in the XML model are hard to diagnose.

2. Some high-level documentation, explaining the XML format, and how it all works, would probably be useful.