Requirement: R2207
Disabled NHibernate 2nd-level cache from hibernate.cfg.xml.  In order for the 2nd level cache to work correctly in a 3-tier scenario, we need a cache implementation that supports clustering, which none of the supplied implementations do.  For now we can work without a 2nd level cache.  If it becomes a performance issue, then we need to find a distributed cache such as NCache or an open-source equivalent if one exists.