Bug: #458 - A short delay is seen before the items in the send/receive queue update initially - A couple of things brought this to my attention: - On startup, the Dicom Explorer seemed to take a long time to initialize. - On opening send/receive queue or Import Activity, a short delay occurred before the items were refreshed.
The fix is quite simple: the Local Data Store activity monitor thread was not always getting 'pulsed' in order to perform a refresh, but instead was only releasing itself on a short 5 second interval. This means that sometimes the delay would appear short, while other times it would appear a bit long.