Bug: #462 Somewhat complicated. Basically the problem is that the DicomClient.Retrieve() method isn't sufficient for the use case where we are using a DICOM server in a separate process. The DicomClient.Retrieve() as it existed in repository creates a listener on the listening port to receive the CSTORE "sub" associations. This is fine if you don't have a server running. When you have a server running, the two processes attempting to create listeners on the same port creates problems. Sometimes, the "sub" association listener gets the CSTORE association first, sometimes the server's listener gets the association first. This is why the observed behaviour seemed random, and you're much more likely to trigger the problem immediately if you retrieve multiple studies at the same time.
The rememdy is to create a new version of Retrieve that assumes that someone else is responsible as a server to receive the CSTORE requests from the other DICOM AE. This has been implemented as a new public method called RetrieveAsServiceClassUserOnly(). The difference is that the network object will be created as a REQUESTOR only. As well, the C-STORE SCP callback is not used.