Bug: #574  UI validation files are not administratively configurable
Changed the paradigm for loading UI validation files.  Instead of automatically loading an embedded resource with the same name as the application component, the validation rules must now be explicitly loaded from a settings class.  The pattern is:
1. Create a settings class with the same name as the application component, eg AddressEditorComponent => AddressEditorComponentSettings.
2. Create an application scoped setting property called ValidationRules of type string.
3. Set the default value of ValidationRules to point to the embedded resource (eg AddressEditorComponent.val.xml).

Although this is far more cumbersome than the original approach, it does allow the validation rules to be changed through the settings editor, which is really the whole point of having them in XML rather than hard-coded.