Requirement: R2609
Modified the API that defines the interaction between the desktop ApplicationComponent and the HTML page.  Now, instead of the app component pushing data to the page at startup and pulling data from the page before saving, the page is responsible for pulling data from the component at load, and pushing data to the component prior to save.  This approach provides more flexibility and a cleaner API.

ris.js: added methods getData(...) and setData(...) which allow the page to get and set data from the component, as per above.
ui.js: fixed an issue where when a row was deleted from a table, the associated error providers elements were not removed.

Modified nuclearmedicine2.htm and breastimaging.htm samples to conform to the new API.