Requirement: R2609
Added methods GetData/SetData to ScriptCallback to allow the HTML page to get/set data from the component.  Modified the interaction between the component and the page.  The component expects to find the following 3 javascript methods defined at global scope on the HTML page:

hasValidationErrors() - returns a boolean indicating if there are any validation errors on the HTML page.
showValidation(show) - show is a boolean flag that instructs the HTML page to show or hide validation errors.
saveData() - informs the HTML page that it should save data.  A typical implementation of this method is as follows:

function saveData()
{
  Ris.setData("mydata", ...);  // send data back to the component
}