Bug: #611  ApplicationComponent.Start - exceptions not propagated to the caller of LaunchXXX()
Modified ApplicationComponentHostShelf and ApplicationComponentHostWorkspace so that the hosted Component is started prior to adding the object to the collection.  This means that any exception thrown from IApplicationComponent.Start() will prevent the Component from launching and the exception will be propagated to the caller of ApplicationComponent.LaunchXXX() as expected.

The only exception to this is with subclasses of PagedComponentContainer (e.g. NavigatorComponentContainer) which are designed to defer starting a hosted component until that component is made visible.  In this case there is no way for an exception thrown in one of the child components' deferred Start() calls to be propagated to the launcher of the container.  In this case, the exception is simply handled by the PagedComponentContainer using the standard ExceptionHandler class.