Bug: #622
This is related to changes made in #611, which cause the Host.DesktopWindow property to be invalid during the call to ApplicationComponent.Start().  Modified IWorkspace and IShelf and removed the Initialize method.  The desktop window must now be passed to the constructor of the Workspace/Shelf.  This should solve the immediate problem.  Ultimately the WorkspaceManager and ShelfManager need to be totally refactored.

-also removed WorkspaceManagerTest.cs, since maintaining it is not worth the trouble