Bug: #630 Often see null reference exceptions from Ris server when loading preview pages
Added 3 methods to the Entity class:
1. Downcast - replacement for C# cast operators
2. Is - replacement for the C# 'is' operator
3. As - replacement for the C# 'as' operator

These methods *must* be used instead of the native C# operators when working with entities (My first thought was to overload the actual C# operators, but it turns out this is not possible).

Outstanding: need to go through all server-side code and make appropriate changes.