Partial commit for Bug: #666.
This is a partial commit.  Updates have been done to several classes to get reading and writing of sequence attributes working properly with the various potential fixed and undefined length encodings that are possible with the DicomWriteOptions.  Note that there was actually a bug in the original source of the library with how it setup the enumerated values for the various types of flags.  The values of the enumerated value were not be forced to be powers of 2, and masks that were being done in the streaming code were causing inconsistent results.