Partial commit for Bug: #545
Added initial managed code DICOM file reader/writer code from the server project.  This will be used to remove references to offis in the local data store to resolve bug #545.  File read/write capabilities were added.  Note also when integrating this code there was a conflict with the class DicomTag.  To resolve the problem, the Dicom\Network\DicomClient code was modified to store tag information as a uint instead of a DicomTag, and the DicomTag class was replaced with the new version of the class.  All references to tags within the imageviewer code now use the DicomTags structure, instead of the constants that used to be in DicomTag with the old implementation.