Bug: #761 Should be a simple way to create custom validation rule
Added a number of validation attributes that can be applied to properties of an ApplicationComponent to perform automatic validation.  This is now the simplest means of adding fixed validation rules to an application component. (Note: does not replace the XML framework that is used to add *configurable* validation rules).

ValidateNotNull - checks that a property is not null.
ValidateRegex - checks that a string property matches a specified regular expression.
ValidateLength - checks that a string property has a length within a specified range.
ValidateGreaterThan - checks that a property value is greater than another property value or a constant reference value.
ValidateLessThan - checks that a property value is less than another property value or a constant reference value.