Commit for Bug: #833 Commit for Bug: #834 Partial Commit for Bug: #835
Changed the file load routines in DicomFile to catch when an error occurs during the reading process.
Changed the default write options to not calculate group lengths.
Changed IO\ChunkStream for Bug: #834, we were not returning the bytes read properly so StreamReader was mistakenly skipping to the end of the stream.
Changed IO\StreamReader for Bug: #833, we now when handling a SQ do the lookup for the DicomTag itself in this routine, and create a private DicomTag object if need be, to avoid the problems in this routine.
Added some minor changes to handle DICOM Asynchronous Window Negotiation.  This doesn't really work yet, although we will pass PDU information with this properly.
Changed NetworkBase for Bug: #835, the Close() method is now in this class and overriden in DicomClient and DicomServer.  We now call it when shutting down a connection.  Also made changes to NetworkBase to handle a timeout when waiting for a connection to close.