Requirement: #734
Added AttributeUtils, a utility class for working with .NET Attributes.  This class has a number of methods for examining types/members to look for the presence of attributes of a specific type, or attributes matching specific criteria.  This class is useful because the built-in facilities for working with attributes are a little harder to use then they ought to be.