Partial commit for Requirement: #678
This is a partially tested/working Query/Retrieve implementation.  Includes a new Storage SCU component for doing DICOM transfers in the ClearCanvas.DicomServices assembly.  Also includes some changes/improvements in the DICOM library to make this functionality work.  The library was incorrectly not sizing PDUs to the max PDU size negotiated, instead it was slightly smaller.