Requirement: #673
Added support for a filesystem monitoring thread that monitors the current status of the configured filesystems.  This also monitors the current used space on filesystems, and selects a filesystem for storing data based on the one with the most space available.

As part of testing, added some better failure recovery into the WorkQueue mechanism.  Added a new failure count to the workQueue, and also added a settings class to configure the number of retries to attempt on failure, and the delay between retries.  Also cleaned up a bit of the error handling when dealing with workqueue entries.

In the future, we may have to debate about how much of this should be in the individual WorkQueue handling code, and how much should be in perhaps a base class that deals with some of the retry issues and setting of status values.