Requirement: #754
1. Integrated Billing/Workload into documentation screen.
2. Rough implementation of saving documentation details.

A number of issues are outstanding, such as the correct timing and validation of details saving, and the fact that billing/workload saving is not part of the same transaction as the details stuff.  These are largely requirements issues - need to correct the design before we can fix the code.