Fix for Bug: #816
Added change so DicomAttributeOW is of type DicomAttributeBinary<byte>.  Also added overrides to SetStreamLength(), GetByteBuffer() and the constructor to properly to byte swapping. for this type.  Also added some regression tests to AttributeTests.cs and FileTest.cs to test big endian and little endian reading and writing of pixel data.