Requirement: #839
Followup to yesterdays commit.  Cleaned up some issues so that the ImageServer service will now run again, and the basic ServiceLock mechanism works now.  This required some code changes and stored procedure cleanup.

Also, updated the GUI so that watermarks can now be configured in the GUI, and added also to the Filesystem table a field with the current percentage full that the filesystem is.  Fixed problem with the default values for watermarks not being correct, and updated the stored procedures for inserting and updating filesystems so they include the watermarks.