Requirement: #1286 EntitySearchCriteria should support querying for a given entity instance
Added methods to EntitySearchCriteria that construct search conditions on the OID, which has the effect of querying for a known instance. Also modified the code generator to use the entity's search criteria class for an entity reference field, rather than the generic ISearchCondition.  So now, in addition to doing this:

Patient somePatient
OrderSearchCriteria orderCriteria = new OrderSearchCriteria();
orderCriteria.Patient.EqualTo(somePatient);

You can also do something like:
orderCriteria.Patient.Name.FamilyName.StartsWith(...);

This takes advantage of HQL's implicit join feature to allow properties of a related entity to be used in the where clause.