Requirement: #1144
Added support for configuring AcceptAnyDevice, AutoInsertDevice, and DefaultRemotePort into the Partition table and to the stored procedures that use the Partition table.  Removed the ImageServerServicesDicomSettings settings class, since this is no longer needed.  Updated the Web GUI to configure these new settings in the ServerPartition dialog.  Also updated the App.Config files to remove the references to the old settings.  Also updated the AssociationVerifier code to now use the Partition table for these settings as opposed to the settings class that was deleted.