Requirement: #928
Initial version of control to view the contents of the ImageServer by partition.  Moved around the master page, and styles so that its more consistent with the rest of the pages.  Also, implemented sorting in the ISelectBroker to aide in sorting the data in the web site.  Still needs some polish and thought on if this is the right implementation.