Requirement: #839
Fixed bug with disk management, where Patient records were never being deleted from the database.  This was caused by problems with the InsertInstance stored procedure, and not maintaining the counts of the number of instances/series/studies in the database, which were depended on to know when to remove a Patient record.  This should be resolved now.