Bug: #1377 Bug: #1378
Fixed problems with ServerRule Add/Edit Dialog.  Made it so the General tab always is open when the dialog first pops up.  Also made it so the XML data is cleared whenever a new rule is being added.  Finally, found another bug and fixed where when adding a new delete rule, the  rule apply time was not set properly.