Bug: #1460
Updates to bring the Shreds implementation in the ImageServer in line with what is being done in the ImageViewer.  Removed the ShredHostExe project and changed the ShredHostService project to be able to be run as both a service and a stand alone executable.  The -service command line option is required for this executable to run as a service.  Also, renamed ShredBase.cs to Shred.cs. Finally, changed the Shreds used by the ImageServer to use Shred.cs as the base instead of WcfShred.cs.  Made some updates to the post build event and the App.config files of ShredHostService.

Finally, ran into some weird problems when testing.  The Commit() on an UpdateContext was not working when no results were found with the WorkQueueProcessor. An exception that the transaction was invalid was thrown.  made a change that we don't commit now if nor esults were found,a nd made some general updates to the Commit() and Dispose() of the UpdateContext to cleanup better and log errors better.