Requirement: #1413
Added new DicomReadOptions option for reading a DICOM file, but skipping over the pixel data.  This helps with performance when trying to just read the header for a file, and there may be tags after the pixel data in the dataset that are to be accessed.  This patch also includes a new GetAttribute() method to DicomAttributeCollection that works slightly different than the indexer for DicomAttributeCollection, in that it will not add a DicomAttribute into the collection, if it doesn't exist within the collection.

This was contributed by Dan Blanchard on the forum here: http://www.clearcanvas.ca/dnn/Community/Forums/tabid/69/forumid/9/postid/1452/view/topic/Default.aspx