Bug: #1473
Cleaned up the network statistics a bit.  Made it so that the StorageScu component now logs statistics.  Made a small change in how the stats are logged when doing averages, so that the individual stats are on their own line, so the log is easier to read.  Also, added a description to the network stats log line so association information is included.