Bug: #1465
Fix for problem where zero length PDU being negotiated, or a near maximum PDU negotiated was causing a failure in the library.  Cleaned up some use of an 'int' instead of a 'uint' to handle the PDU length.  Also made changes to AssociationParameters so that we now keep around the local and remote Maximum PDU size.  Finally, changed how the max pdu size is set when creating PDUs so we use the local max pdu size if it's less than the remote max pdu size.

Also removed the TransferMonitor class, this is no longer being used, and made some Resharper related cleanup to the code.