Bug: #1481
Updated to the 2008 version of the DICOM standaard.  The Data dictionary was updated.  The DataDictionaryGenerator project was updated to work with the new XML version of part 6 of DICOM, and a few updates were made to the code where tag names changed to properly reference the right tag or SOP Class.  Finally, the ImageServer default data script was updated to include the 4 new Storage SOP Classes that were introduced into the standard.