Bug: #1322
Updated the AddEditDeviceDialog and AddEditPartitionDialog screens to enable/disable text box entry forms in client side javascript.  This was linked to the IpAddress field on the Device screen when DHCP was checked, and the Default Remote Port text field on the Partition dialog when the AutoInsertDevice check box was not checked.  Also, cleaned out some unneeded javascript in the partition page, added a new column on the partition grid view, and changed the device grid view to not display ip address when DHCP is checked.  Also, did a small amount of code cleanup of the javascript generated by the RangeValidator.  No functional changes in this, however.  Finally, changed the number of rows displayed on some of the screen so they fill up the grid portion of the screen better.