Update Requirement #1296: Add Validation for Filesystem Paths for ImageServer Web GUI
This commit addresses the issue with the validation always fails when a network folder is entered. It has to do with ASP.NET security. ASP.NET process by default run under ASPNET account which has restricted access.

To access to the folder, we need impersonation enabled in the web config. The user credential should be set during the installation. (see <identity> in web.config). However, it seems that any threads spawned within an ASP.NET web service will always run under default ASPNET account despite the impersonation in enabled in the web configuration. Therefore, the thread is not used for now until we figure out how to force it to run under a different account (a post has been entered on ASP.NET forum but there's no answer so far).