Bug: #1228 Bug: #288
Patched log4net to include a new MutexLock locking mode.  It can be configured from the Logging.config file like this:

	<lockingModel type="log4net.Appender.FileAppender+MutexLock" />

This new mode keeps the log file open, and uses a named Mutex to lock the file while it is being written to.  The ReadMe.txt in the ReferencedAssemblies/log4net directory describes how the patch was done, in case we have to rebuild it in the future.  Also made a small update so Shred threads are named so they show up named in the log files.