Bug: #1505
Resolved issue where invalid study date causes XML processing problems and Q/R problems.  First, made a change so that if we have an exception setting a tag when doing Q/R, we catch the exception and set the tag to null, instead of failing the query.  Second, added a new settings class to the DICOM library to enable/disable validation when setting tags.  Made validation be disabled by default for values of tags, but enabled for lengths of tags.