Requirement: #1130
Final commit for implementing codec interfaces.  Added documentation to the code and fixed some minor bugs.  Specifically, fixed a problem where 8 bit data was not having its StreamLength calculated properly.  Also Added a check to the code with the rle decompressor to prevent an array index out of bounds exception on some rle data.  it looks like some rle compressors are putting a 128 instead of 0 for padding of segments.