Requirement: #1545
Added schema validation to the ClearCanvas.Common.Specifications and ClearCanvas.Common.Actions namespaces.  All of the operators and actions now define their own XMLSchema and the compilers take these schema elements and bring them together into an XmlSchema.  The Compile() option of the compilers can now specify if they want to do schema validation or not.  Note that the constructor creates the schema no matter what.