Requirement #1512: Add new Study/Series Details screens to web GUI
Added Series detail page. The page is accessible by double-clicking on the series in the study page or clicking the open series button.

Added SectionPanel template control for consistent UI look. Each section consists of the heading and the content which can be anything. In this case, the contents are the study and series info. The heading and the content area can be styled separately.