Requirement: #967 XML Validation
Fixed a bug where the Scintilla control will not load because it cannot find the native SciLexer.dll, resulting in a "Window class not valid" error.  The fix was relatively easily as the control has a constructor overload that allows you to pass in the path to the native dll.  Note that a CommonDirectory property was added to the Platform class which returns the fully qualified path to the "common" folder.  The SciLexer.dll is deployed into this directory.