Bug: #1704
The way the TransferSyntax and SopClass classes were initializing their static list of transfer syntaxes and sop classes was not thread safe.  This could result in unknown sop class/ transfer syntax messages if multiple files were being read at the same time.  Modified the classes to use a static constructor to init the dictionaries, which will be thread safe, and also thus does not require the use of a lock.