Requirement: #1399
Introduced a new ServiceLock service for reapplying rules to online studies.  The service is disabled by default.  The service, however, is inserted into the servicelock table whenever a filesystem is added.   The service scans a filesystems for all studies, and then reapplies the rules engine and currently updates the FilesystemQueue entries for the study.