Requirement: #1145
Added support for doing DICOM move requests from the web GUI.  The commit is functional, but it is an interim commit and needs some polish.  The move is launched from the search screen, a new page is displayed listing the studies requested for moving, and a device is selected for a destination.