Bug: #1785
Fixed default setting in app.config for the number of WorkQueue threads, this should resolve the problems with the server keeping up with incoming processing during testing.  Also introduced new settings related to the ServiceLock table for the delay until rechecking a filesystem after a delete and also for the delay inbetween checks when below the watermark.