Bug: #1885 and Bug: #1886
Completed adding new TimedDialog dialog for a modal dialog that can timeout after a configured amount of seconds and be dismissed.  A confirmation event is added to allow redirects to a new page or other actions when the dialog is dismissed, or the user hits an option "Ok" button.  These tickets are complete.