Requirement: #1626
Added new device column for allowing auto-routing.  Also changed the auto route rules action item to check if a device has been enabled for auto-routing before we actually insert a route request.

The changes still need a new icon to represent this functionality for the Device screen.  A placeholder has been created for the icon.