Bug: #1888
Cleaned up how deletion of WorkQueue entries occurs.  We now have a stored procedure, DeleteWorkQueue, which does this cleanup.  It locks the study before doing the delete so that the study will not be processed when the delete is occurring.  Also, added new CleanupStudy WorkQueue entry for cleaning up failed workqueue entries and the corresponding filesystem.