Requirement: #849 Requirement: #1906
Added policy mechanism for duplicate SOP classes.  There now is a mechanism to select a duplicate SOP policy by server partition.  A new enumerated table has been added, DuplicateSopPolicyEnum, where the policies are specified.  This table is linked to the ServerPartition table.  The user can now select the policy in the partition add/edit dialog, and it is displayed in the grid view for partitions.  Code changes were made in the dicom service and the work queue to implement the specific policies.  Also cleaned up some issues with dealing with failed objects from a previous checkin during testing.