Bug: #1960
Changed how duplicate images are reprocessed.  Now we will just fail the queue entry if all entries are duplicates when processing, instead of going into a retry mechanism.  Note that there might be a case where 1 retry is done is some images are true duplicates within a study and others are different.