Bug: #2019
Several users were having problems with the server not receiving images w/ group 0x0004 tags in them. Even though this is technically illegal dicom, we need to deal with it.  Changed the calls to the DicomAttributeCollection constructor where we were setting tag 0x0008000 as the start tag to be tag 0x0004000 instead.