Defect #1925: Processor ID on Schedule page may not display the most useful IP
This is interim commit. More discussion is needed to see if this solution is acceptable or a more general solution should be used because the RIS may also need to manage servers within a server farm. The approach used in this ticket is meant to facility the communication from the web server to servers in the future, which makes it fall under the same umbrella.

Changes in this commit:

- Added ServerInformation table to store the information about servers in the cluster. This table contains the HostName, ip addresses of the server and the a timestamp indicating when was the last time the server was "active".

- Added foreign key to WorkQueue table to the ServerInformation table. This key is populated by the work queue processor on the server.

- Added "Processing Server" fields in the work queue result screen and work queue detail page. These fields indicate the hostname of the server that is processing the entries.


Note:  For now, a shred on the server registers the server information when the server starts, and periodically updates the ServerInformation table to indicate it's still alive. At this point, the web server cannot communicate with the wcf service of the shredhost on the server (which is the desired way to control or query live status of the services) because the wcf config (eg endpoint and binding) is not available outside the ShredHost.

If this design is to stay, further discussion needs to address how the server contact information can be obtained automatically unless we want users to enter the information manually from the GUI. (can we add an extension to the ShredHost to publish it?).

Note:  The server information for the ServiceLock is not implemented as part of this commit, pending on the outcome of the discussion.