Requirement: #2174 Order notes posted to me folder should be split
In order to allow folders to be added to folder systems dynamically, some preliminary refactoring of folder systems code was needed in order to simplify and make it possible for humans to work with it again.  This commit is a first-pass attempt at this.  Mountains of duplicated code were deleted by simply moving stuff into the base class (that felt good!).  More to come.