Bug: #2095
Finally patched this from the 0.95 release code of the ImageServer.  Whereas there we just forced regex to no longer match on null values, in this implementation I've added a new optional attribute to the <regex> operator, "nullMatches", which is a boolean telling if null should match, with a default value of "false", that it doesn't match.