Requirement: #1907
Fixed problem with MoveScu not shutting down properly on some move status values.  Did some general cleanup of the FindScu and MoveScu based on resharper suggestions.  Made change to BlockingThreadPool, such that we catch exceptions if they occur within the thread, log a message, and properly cleanup the active counts.