Bug: #2292
Modified the servicelock and workqueue processors to use the new ItemProcessingThreadPool class for processing queue items.  Also, modified the checks on how many threads and items are in the thread pool queue so that we only queue as many items as there are threads (and thus not overqueue).  Finally, made some adjustments on how many objects are processed when processing a queue item according to the priority.  This was based on anecdotal testing with the migration tool (and seeing the server take too long to process some studies), and is only slightly related to this ticket.