Requirement: #1949
Added retry mechanism when services startup and the database is unavailable.  (Note that these changes do not impact when the service is already running and the DB goes down.)  How services are managed/started up has been modified so there's a init phase and then a run phase.  We go into a retry loop when an exception happens in the init phase.  We re-try the init every 60 seconds until the service is shutdown, or the init succeeds.  As part of this some of the WorkQueueProcessor and ServiceLockProcessor code was refactored.