Bug: #1084 & Requirement: #1741
Reworked the DICOM services such that we now dynamically start/stop DICOM listeners based on updates in the database.  We now poll every 60 seconds for changes to the ServerPartition table, and start/stop and update partition information on these intervals.  With these changes, the dialog in the GUI to restart the service has been removed.