Requirement: #2389 Add support for separators to action model
1. Refactored the Path class to have immutable semantics, and added a few methods to public API to make it more useful.

2. Added support for separators to the action model.  Separators can only be specified via XML.  The syntax is:

<action id="x" path="A/B/C" />
<separator/>
<action id="y" path="A/B/D" />
<separator/>
<action id="z" path="A/E/F" />

The separator location is determined by taking the longest common path of the immediately preceding and succeeding action elements. In this example, the first separator is placed between C and D, and the second between B and E.  The current algorithm has the limitation that if action x or action y does not exist (e.g. its a workspace tool, but that kind of workspace is not currently active), then separator 1 will be ignored.  Not sure if there is a way to overcome this limitation.

3. Also modified ActionModelSettings so that changes made in the settings manager take effect when saved, without having to restart the app.