Ticket: #1910 Ticket: #1911
Added support for inserting into the ArchiveQueue when a new study is processed that will not be deleted (ie, it doesn't have a delete rule).  Also, implemented support for inserting into the ArchiveQueue if required when reprocessing studies (and reapplying rules).

Also, refactored the ServerCommandProcessor related classes into a subnamespace in ClearCanvas.ImageServer.Common.  Removed two stored procedures, QueryWorkQueueUids and QueryRequestAttributes, which could just as easily be done using the entity brokers, as part of the database work.  Finally, default values for the new enumerated value tables were not checked in before by mistake.