Ticket: #1910 Ticket: #1911
More updates to the HSM archiving code.  Removed the SharpZipLib, and instead added use of the DotNetZip, which has better licensing.  This library can disable compression in the zip file, so in reality, its probably better for our uses than tar from SharpZipLib.

Also, cleaned up the code and added some documentation.  The archiving code is now functional, and we also now have a shell for doing restores.  The archiving code uses the ServerCommandProcessor to do the actual archiving, and will fully rollback its work if a failure happens during archival.