Bug: #2409 problems with homepage search
Changed behaviour of Search Results folders such that they can be refreshed as any other folder.  However, unlike workflow folders they do not support updating the count independently of the contents.  Hitting refresh will update the search results if that folder is currently selected, otherwise it simply invalidates the folder, and it will be updated the next time the user selects it.