Bug: #2094
Increased the length of the FailureDescription column from 256 bytes to 512 bytes. Also as part of this change, decided to decrease the length of ProcessorId from 256 to 128 to help reduce the row size for the WorkQueue table.  (Since we're currently just storing ip addresses in this column, its even longer than need be with the change to 128).