Bug: #2779
Created two WorkQueue processors.  Added functionality to the WorkQueueProcessor class, where you can specify a list of WorkQueueTypeEnum values that the processor should handle.  Changed the WorkQueue settings class to be updatable (this wasn't really needed), and also created new Primary / Secondary work queue configuration values for setting the thread count, and then an array of workqueuetypes to be processed by the Primary and Secondary processors.  A comma separated list is used with the "lookup" values for the enumerated value.  Also fixed a small build problem with the codecs not being put in the plugin directory properly.