Requirement ##840 : Verify Demographic Data of Incoming C-STOREs
This is initial commit based on a new design. A StudyHistory table is added to keep the history of the patient demographics. It contains different sets of demographics that were mapped to a study through reconciliation. New images can be reconciled automatically if a history record already exists with the same demographics. Otherwise, the images will be put into the ReconcileQueue. The StudyHistory will be populated by a GUI when user reconciles the images.

This commit contains preliminary changes to the StudyProcessItemProcessor to query the StudyHistory for historical reconcialiations and insert new ReconcileQueue when it doesn't find any.