Requirement: #2065
I think this should close this one out.  I believe the comments in the ticket about the failure description not showing up are related to the fact that we previously only displayed the failure description when the queue entry had a "failed" status.  We now will display the failure description no matter what the status, if it exists in the queue entry.