Bug: #2938
Fixed parsing bug in code.  There was a specific encoding where a UN VR tag was the last value in a sequence of defined length, and the length of the UN tag was < 4 bytes where some code that we used to determine if a UN tag was in reality a SQ attribute was mistakenly going past the end of the SQ tag, and thus causing an "end of stream" exception to be thrown.  There was a general problem that we weren't calculating how many bytes were needed in the stream before we did the read which is now resolved.  There was also some additional code that we had commented out before that was removed to help with this.