Bug: #2938
Continuation from last checkin.  Decided to slightly change the parser to not ask for more data if it doesn't exist when trying to figure out if a UN VR tag is an SQ.  Since this is just a "guess", doesn't make sense to return here asking for more data if the check can't be done.