Bug: #2682
Went through all the code where the storage location was being loaded, and made sure it was properly checking if the filesystem was readable.  Created new method for getting the storage location in FilesystemMonitor.  Changed the code to call this routine in most places to get the location.  Note that these changes backed out some of the code that handled when there was more than one filesystem storage location.  Note that we currently not using that functionality, although we might add it in the future.  Also did a bit of cleanup of the code with these changes, including some changes to the header streaming code and how it is organized.