Bug: #2653
Implemented both memory and performance optimizations for working with the StudyXml.  In some simple testing with the ImageServer, memory usage is at least cut in half,and performance is more than doubled.  For speed improvements, after profiling discovered the DicomAttributeMultiValueText.Equals() method was innefficient, made improvements with this.  Also discovered that how we were retrieving tags from a DicomAttributeCollection via the Contains() method and the indexer were innefficient.  Implemented a new DicomAttributeCollection.TryGetAttribute method which combines the functionality in these methods and reduces the number of times when working the StudyXml that the tag must be found in the collection.  This should lead to some general performance improvements in all apps using the library.

For memory performance, took advantage of some code that Stewart introduced for not parsing the an Instances Xml data until its needed.  Now, when we write the StudyXml out, we cache the Xml data.  When this occurs, we now free the DicomAttributeCollection associated with the InstanceXml.  If the collection is needed again, its reconstructed from the cached xml data and the base instance.  This accounted for more than a 50% drop in memory usage in some basic tests.

As part of this, also wrote a new regression test for testing the study xml functionality which creates a study xml file, reads it back in, and compares with the source attribute collections to make sure the tags are the same.