Bug: #2861
Added "READPAST" locking hints for queries against the ServiceLock, WorkQueue, ArchiveQueue, RestoreQueue and FilesystemQueue tables where we're looking for work to do.  This should allow us to just skip over any rows that may be locked when looking for a row to work on.  The MS documentation specifically mentions this as useful for "work queue" type tables.