Bug: #3024
Reworked the WorkQueue scheduler and WorkQueue related configuration options.  We now have a bit more sophisticated mechanism for scheduling queue items. A total thread pool size is configurable, along with the number of threads that should be dedicated to high priority queue entries and the number of threads that should be dedicated to memory intensive queue items.  The remainder of the threads can be used for non-memory intensive queue items.  We now first try and schedule the configured number of high priority queue items, then we take the next scheduled queue item until the number of memory intensive queue items is maxed out, or until the thread pool is used up.  If the max number of high memory queue items is reached, then we only look for non high memory usage queue items.