Bug: #2977
Updates to the codecs so they now set lossy image compression and derivation description properly.  Also, found a bug where the J2K compressor was never doing lossy compression, it was always defaulting to lossless due to a small issue with when we were setting the transfer syntax in the output DicomCompressedPixelData object (we were waiting until after the data was compressed instead of before).