Bug: #3047
Fixed the problem by adding exception handling to when the file is loaded.  Also, added a loop if we fail loading one file in the study, we go forward to the next file and attempt to load it.  Note this code will still force the "bad" file to get into the WorkQueue, so that it can be processed, or deleted/cleaned up if necessary.