Bug: #3068, Bug: #3069, Bug: #3049
A bunch of cleanup of restores.  Fixed first two bugs related to the study state in StudyStorage not being correct, and with the rules engine not being applied after a restore.  Also, changed restores such that we can now restore any archived study, and the restore will occur over the top of the current study.