Bug: #2686
Introduced a new Restoring status to the Restore Queue, which is used after a study has been read once and failed the read.  Also added a new configuration option that specifies how many simultaneous restores happen at once.  Also made an unrelated change to include the study instance uid in the data stored in the database for each study archived, and we now use this uid when restoring to determine the directory of the study to restore.