Ticket: #3223 Feedback of image availability in the RIS client
Renamed IImageAvailabilityUpdateStrategy to IImageAvailabilityStrategy and refactored it so as to shift the balance of responsibility between the ImageAvailabilityWorkQueueItemProcessor and the strategy.  The strategy is now purely computational - it has the sole responsibility of computing the image availability for a given procedure, but it does not update the model.  The processor is now entirely responsible for creating and updating work queue items.