Ticket: #3472 Orders are overly susceptible to concurrent modification
Turns out this is mostly due to the Order.Notes collection which is modified by many different parts of the workflow, because it maintains different "conversations".  Temporary solution was to remove this collection from the Order entity, so that the insertion of new notes does not marked the order as being updated.