Ticket: #3518 DOB is off by a month in the previews for months with less than 31 days
Modified javascript parseISOString function to use the new Date(y, m, d, h, n, s) constructor, rather than setting each property individually, which would yield odd side effects when the day was already set to a number larger than the number of days in the month m.  It is not clear why this constructor was not used in the first place.