Bug: #3693
Changed calculation of free space by the FileSystemMonitor class so that it isn't done within the lock.  Also made a small performance improvement in some of the web data sources to search based on study instance UID and server partition GUID.  This was technically a bug, but it should improve query performance.  Also, changed how often we check filesystem usage to every 2 minutes instead of every minute.  Also, a minor cleanup on the database default table sizes to improve performance.