Ticket: #3488 Print report code may have memory leaks
Instead of creating a new instance of the browser for each copy of the report to be printed, the component now renders each report in the existing preview window and prints from this window.  This avoids creating multiple browser instances, which in the old code, were not being cleaned up.  Also applied the same fix to the downtime printing code.