Ticket: #3743
Completed commit.  Added cancellation / shutdown support to the ServiceLock processor. Similar to the WorkQueue code changes, introduced a new ServiceLockThreadPool, which keeps track of the service lock entries, and also cancels them when appropriate.  Also introduced into each of the ServiceLock processors the ability to cancel in the middle.  Note that when a reinventory or reprocess is canceled, both of these operations reschedule themselves and start up again running the entire reinventory/reprocess.