Ticket: #2334
Partial commit of updates to handle DB upgrades.  Made changes to the code/model generator to now have a constructor that takes the columns in the model as parameters to each model entity.  Also added a static insert routine to each model where we now can insert a new row based on a model object being populated.  Also, added the DatabaseVersion_ table into the creation scripts.