Ticket  #3557: Add StudyHistory tab to the Study Details page
Interim commit. Added a history tab in the study details page to display the edit and reconciliation history (which result in updating the study). When possible, the information will be decoded to make it readable.

Also fixed a bug where Study Process entry fails when the study is edited and then new images with different demographics is received. When processing new images, we should only look for reconciliation record in the Study history table.