Ticket: #3835
Changed the rules engine so it is run after a study has been processed (when the WorkQueue entry is being deleted), instead of when it initially is processed.  This also should cause FilesystemQueue entries for late arriving images to be updated.  We also apply the rules to the first image in each series in a study.  Added changes to make the rules be applied when a CleanupStudy WorkQueue is called.  This will cause us to have the rules be run on a study, even when its initial WorkQueue StudyProcess entry failed.  Also made some updates to how PostProcessing() is called for all the WorkQueue entries.