Ticket: 3641
Added support for purging old alerts.  Generalized the log writer that was used for the application log, so that it also could be used for writing out alerts.  Added new ServiceLock to do the actual purging of alerts.  Also, added config options in the ServiceLock section related to setting up purging/archival.