Ticket: #3864 - Ability to throttle max connections to a device
Interim commit.

This commit includes changes to the database and auto-route/study move work queue processors. When the connection limit on a device is set, auto-route and study move entries will be delayed until some of them are completed. A app config setting is also added to indicate if an alert should be raised when the throttle setting is not set and the number of connections appears too high.