Ticket: #3864 - Ability to throttle max connections to a device
Added new tab in the device configuration dialog for entering the throttle settings. The default max # of connections for new devices is specified in the app config.

As part of the change, refactored the validator classes so that it's now possible to associate a validator to a checkbox/radio button.