Ticket: #2334
Updates to the upgrade mechanism based on feedback from Jonathan, there was some poor design decisions with the previous rev of the tools.  We now no longer have the last script having the destination version being null, its not set to the assembly version of the assembly where the upgrade scripts are located.  Also cleaned up some interfaces, and added some comments.