Requirement: #4225
Significant cleanup of logging from the migration tool.  Reworked so just information about the actual studies being migrated and failure conditions are now logged to make it a bit easier to track what is happening.  This included some changes to the SCU classes in the DICOM library to allow disabling of some informational level logging.  Also made general updates to the tool to reflect updates in the framework.