Ticket: #3407 Add Audit trail logging to Web UI.
Added infrastructure for IUserCredentialsProvider as discussed in emails, and added a UserCredentialsProviderClass setting to the RemoteCoreServiceProviderSettings class, which allows specifying the credentials provider as a class name in the configuration.

These changes should not break any existing code, because the credentials provider is treated as optional - if none is provided, credentials are not used.

TODO: create some standard implementations of IUserCredentialsProvider, if possible.