Bug: #4225
Updates to the Migration Tool.  Found a bug in the general DICOM library where we could exit out of the processing thread for a SCU association,and not signal the main thread we're done, which would cause the main thread to lock.  Lengthened the MoveTimeout for move associations to be longer.  Also found some small cleanup issues.  Finally, made some improvements to the gui for the migration tool to show the count of results on the main search screens.