Ticket: #4160 Create a Folder Explorer configuration page in Tools/Options
- fix a bug where the folder tree rebuild and flicker when moving folder system up/down/dragged, etc
- fix a bug where the folder system cannot be dragged to the end of the list (empty area)
- fix the weird border around the folder tree icons

Note: The CheckStateChained behaviour cannot be changed so that when a node is checked, all children nodes are checked.  This is because when a child node is checked, it check the parent, which then check all the siblings of the first checked child node.  The whole chained logic needs to be reviewed, especially when involving drag & drop, etc.