Bug: #4263
Reworked how the Writeable/Readable flags work in the GUI, and how they relate to the services.  Previously, in most of the service, we were doing a Readable check when looking for the location of a study, now its a check to make sure the study is "Online".  Also found a case where the Filesystem Watermark Check was not running when a filesystem was Write Only.  With these changes, the WriteOnly mechanism basically doesn't do much anymore.  Data can still be read from the filesystem.