Bug: #4299
Changed how we query for rows in the application log.  We now first find the oldest log, and then if it should be archived, we query in 5 minute increments until we're past the end of the logs to archive.  Hopefully the smaller chunks of rows will help eliminate the previous problem.  Also, change the default number of days archived to 1, so that only the previous day's logs are kept around.