Requirement: #4133 - Rework Duplicate Sop Policies
Interim Commit.

This is first shot at the requirement. New study integrity queue type is added for duplicate sop entries. Additional columns are added to the StudyIntegrityQueueUid table to store the information about the source and destination of the transmission. These will be used together with the demographics in the image to group duplicate sops transmitted through different associations to the same or different servers.

Duplicate sop instances received through Dicom service or imported from filesystems are handled by the DuplicateSopHandler, which inserts/updates a record in the StudyIntegrityQueue table and copies the duplicate into the "Duplicate" folder.

A new column is added to the Study Integrity Queue Search screen to identify different types of entries. A new dialog box for duplicate sop reconciliation has been added but the action buttons are disabled at this point since the back-end hasn't been implemented.

The test app is updated and can be used to send duplicate sop instances simultaneously to the same server.