Bug: #4419
Fixed problem with ArchiveQueue/RestoreQueue not recovering after a crash.  We did not have a mechanism when the service came up, to look for In Progress entries in the queue, similar to how this is done with the WorkQueue and the ServiceLock processors.  We now will check for In Progress entries in these queues when the archive service comes up.  The entries have to be tied to the specific process, and archive when its starting up.