Bug: #4196
More changes to improve query performance.  Traced the performance issues on the migration server back to having to check the study status/queue study state on every row in the found set.  Added the StudyStorageGUID as a column in the Study table, and then added the ability to do EXISTS/NOT EXISTS selects from the study table against the StudyStorage table.  The query now queries the online objects, then queries all the nearline objects when a dicom query occurs, putting the results together and sending them all back.  This should improve performance.

Also made updates as part of this to support NotIn clauses in the ISearchCondition.   This will require some updates to the Hibernate wrappers still to support.