Bug: #4454 Bug: #4179
Another go around with changes in the StudyXml generation.  Changed how the settings for the StudyXml creation is done, we now set if each type of tag is included, if its treated as excluded, or just ignored.  Also, changed the Copy() methods for the DicomAttributeCollection so that we can specify if Unknown & Private Sops should be copied, this should reduce our memory usage with large studies, and hopefully improve our processing time, and also potentially address the IO related exceptions being experienced.