Bug: #4372   Duplicates received while auto patient name reonciliation occurs for previously received images can lead to incorrect instance count
Could not reproduce the problem described in this ticket.  This commit contains changes that address another issue that was observed: work queue failures if duplicates arrives while auto-reconciliation is taking places.

A check is added to the DicomSopImporter to ensure there's not another copy of the sop instance in the reconcile folders (either for manual or auto-reconciliation). If another copy already exists, it will reject the sop instance and returns "Duplicate Sop" error to the caller (dicom service or file importer).

StudyProcess processor is also modified to handle the case where the original files in the study folder which caused the dicom service to think a duplicate was received no longer exists (due to auto-conciliation). When the original file is missing, the duplicate will be thrown into the SIQ queue as a duplicate entry immediately. Otherwise, it will be compared against the existing instance and deleted if it is an exact match.