Bug: #4275
Did some further investigation on the FileStream class, and discovered another form of the constructor, where the FileOptions could be set, and specifically where a "WriteThrough" option could be set, that disables caching when writing, and forces the data to be written directly to disk.  Made the use of this flag when writing to a file configurable, but have the option enabled by default.  I believe this change should address the issue with files being written that were the write side, but they were blank (all nulls).  The assumption is that when writing, these files were cached, and just hadn't been written out when the power was drawn.