Bug: #4179
First attempt/pass at re-fixing this.  Still not completely fixed, but its operational, and works better than before.  THere's a bug that after saving an XML file once, and then the file is reloaded and rewritten out, the excluded tags appear again in both the base instance and the child instances.  Works correctly, however, on the first write out of the file.