Ticket: #4612 Add failover mechanism to RemoteCoreServiceProvider
Improved performance of the failover mechanism by building in a little bit of intelligence.  Now, when a node (either primary or failover) is unreachable, it is marked as being "blacked out" for a period of 30 seconds.  The client will not attempt to contact a blacked out node until the blackout period expires.  This prevents the client from wasting time continually trying to contact a node that is down, which completely destroys the user experience.  A consequence is that, when a primary node comes back online, the client will may not switch back to the primary node for up to 30 seconds.