Bug: #4658
Fixed problem with the WorkQueue not showing entries in the web gui that had not been inserted into the Study table yet.  Although not confirmed, when doing a code review found a similar issue with the ArchiveQueue and RestoreQueue and Alert page.  Although it wouldn't cause missing records, some queries added a wildcard character on a search for some values that wasn't needed.