Bug: #4677 Reprocessing study should keep the history and entries in the reconcile queue
Changed ReprocessStudyItemProcessor to use SopInstanceProcessor to process images instead of spawning a ProcessStudy entry. Any files that are not readable will be deleted and alerts will be generated. The object count in the database will be updated to reflect what's in the study xml.

Added new type of StudyHistoryTypeEnum to log history of reprocessing. Moved the code to insert the ReprocessStudy entry into StudyReprocessor class (will need a better naming convention to differentiate classes that insert the work queue entries and those that actual process the entries).