Ticket: #4523 Develop Auto-Recovery Mechanism for WorkQueue Failures
Fixed problem with the Reprocess button in the GUI not enabled when there's item in the work queue (eventhough it already failed). Users should be able to reprocess the study anytime it's idle or all the work queue entries have failed.

Also added new authority token for the study reprocessing.