Ticket(s): #4804 - Fix for Instance Availability
Schema/Configuration Changes: N/A

When the Instance Availability tag was set to NULL in the C-FIND-RQ, we were mistakenly setting the value to null in the C-FIND-RSP (overwriting the actual value we had set).  Technically the value shouldn't be set in the RQ according to dicom, but applying this fix to make the code still work when its included.