Ticket(s): #4807 - Uncaught exceptions when processing duplicates
Schema/Configuration Changes: N/A

This was alittle more widespread of a problem than reported in the ticket.  When the Import/Incoming folder functionality was added, some shared code was produced to handle both cases.  The shared code introduced throwing an exception in several error conditions.  The code calling the shred code was expecting a failure return status instead.  This was causing associations to be aborted, instead of just sending a failure status when these exceptions would occur, which is somewhat bad DICOM.