Ticket(s): #4807 - Uncaught exceptions when processing duplicates
Schema/Configuration Changes: N/A

On checkin, noticed some further changes that should be done with the DuplicateSopProcessor class, and how it handles errors.  Made it return the error status instead of throwing an exception, and then all the places it was being called, added a check of the results from the processing.