Ticket: $3803
Added File System filter to the search parameters. Each of the file systems can be filtered. Since the descriptions are used, it is possible to have duplicate or more descriptions in the filter, however the GUID of the filesystem is used to search, so that specific file system will be returned.