Ticket: #4968 DDL writer naming conventions are not very consistent
Manually merged changes back into Trunk.

1. Created a custom algorithm to generate names for indexes, unique and foreign keys, rather than using the names that Hibernate generates.  These names include an MD5 hash component to help ensure uniqueness.  All names now follow the format PAM where:

P is a prefix { UQ_ = unique key, FK_ = foreign key, IX_ = index }
A is the human readable component of the object name, consisting of either the concatenated column names covered by the object, or a descriptive alias.
M is the 32-character hex message digest of the table name to which the object belongs plus the content of A.

Note that A may be truncated because the entire name is constrained to 64 characters.

2. Explicitly model unique columns as unique constraints, so that all unique constraints are now treated equally.  This simplifies the Renderer code as there are no longer any special cases to consider.

3. Changed all unique-key names in hbm.xml files to end with an underscore, to follow our naming conventions for other DB objects.