Ticket: #4761 - Add ability to delete series from the series tab of the study details page
Fixed problem in the scheduling manual reconciliation for new images. Changes in rev 10780 added Source AE Title to the SeriesInformation class so that it can be displayed in the study history. The SourceAE is extracted from the Series table. However it is not available in the context of the manual reconciliation scheduling code. The easiest fix is to remove it.

Schema/Configuration Changes: N/A