Ticket(s): #5135 - Add db Matrix that sets the characteristics of WorkQueue items
Schema/Configuration Changes: N/A

Introduced new table, WorkQueueTypeProperties, that holds the characteristics of WorkQueue entries.  This includes the default priority, various delays, number of batch records to process, and if a WorkQueue item is memory limited.  Updated query related stored procedures to now use this table and simplified other parts of the code based on this new table.  Also added default configuration for the table.  This also includes changes to pull out some of the WorkQueue related configuration options.

[t:5135]