Ticket: #5231 - Error when deleteing multiple studies.
When deleting, the partitions for each of the deleted studies are refreshed. So, if two studies are deleted from the same partition, that partition will be refreshed twice. Changed the code to check if the partition is set to be refreshed already, and to only refresh each effected partition once.