Ticket: #4874 Attempt to insert a ProcedureTypeGroups with duplicate name does not generate a user-friendly error message in all cases
Schema/Configuration Changes: N/A

Infrastructural changes made in previous commits related to this ticket required that the EnumValue class be updated.  Previously, Unique attributes were defined on the Value and Display properties of the EnumValue class.  However, this is not logically correct, as the constraints do not apply across the entire set of EnumValues, but rather, only to a distinct subclass of EnumValue.  Therefore, the constraints need to be expressed on the subclass.  The only way to do this is to use the UniqueKeyAttribute on the subclass, instead of the UniqueAttribute on the properties.

1. Removed UniqueAttribute usages from EnumValue Value and Display properties.
2. Modified codegen/templates/EnumValue.gen.ct to apply UniqueKeyAttribute to the generated subclass, on both Value and Display properties.
3. Fixed a bug in the definition of UniqueKeyAttribute - multiple instances are allowed.