Ticket(s): #4982 - ConvertColorspaceToRGB doesn't work properly
Schema/Configuration Changes: N/A

As per the comments in the ticket, the dinfo.jpeg_color_space line needed to be removed.  This was forcing the decompresser to assume the source data was YBR encoded, when it was actually RGB for lossless, and causing an incorrect decompression.  We now force the ConvertColorspaceToRGB codec parameter now works properly, and is set to true where appropriate.

Note, in testing this ticket, the comments as to how ticket #4974 was tested should be redone with this ticket.
[t:4982]