Ticket: #4984 - Rules engine not applied when restoring lossy compressed study
Schema/Configuration Changes: N/A

Added applying of the rules engine on a restore of a nearline study.  Found a couple additional defects while working on.  One, there was a null reference exception on a nearline restore.  Two, there was an issue with the link created on the Studies page to the RestoreQueue and the ArchiveQueue.  These both were actually going to the WOrkQueue page, instead of the RestoreQueue and ArchiveQueue when the state was restoring/archiving.  Also found issues with the encoding of the URL when referencing the other pages.  Had to add explcit UrlEncode/UrlDecode of the patient name and patient id.

[t:4984]