Ticket(s): #4995 - See errors in log on production related to reconcile committing of transactions
Schema/Configuration Changes: N/A

The BaseAdaptor.Delete() method had a inadvertent Commit() in it, which caused the issue.  In looking for usage of this method, also discovered that the same error occurred when deleting a RestoreQueue entry from the web gui.  The fix fixed this also.  Finally, did some resharper related cleanup with the ticket, and also cleaned up a warning compilation message at the same time.