Ticket: #5283 - Late arriving image for study on write disabled file system cannot be processed
Schema/Configuration Changes: N/A

If a study is stored on a readonly filesystem, and new images arrive, the software will now look for the incoming folders specified in the ServiceLock table.  If one exists for a "writeable" filesystem, the DICOM service will now store the data to this folder and send a success to the client.  The entries will then stay in the Incoming folder until the filesystem comes back online and is writeable.