Ticket: #4767 - Image Server does not handle images properly if the patient's name tag exceeds 64 characters
Schema/Configuration Changes: N/A

Interim commit

Modified the code to trim space characters from the patient's name when processing new images (on the live system, the sender appended spaces to patient's last name to
make it 64 character long). The option to remove these redundant spaces is configurable in the app config. For existing SIQ entries that suffer from this issue, the SIQ GUI code has been updated to remove the spaces in the patient names.

Updated the look-and-feel of the reconcile history record in the study history tab.

Added unit test for DicomNameUtils and moved some of the preprocessing code in the Study Process processor to the core.