Ticket: #4163 - MoveScpExtension does not dispose the ImageServerStorageScu object
Schema/Configuration Changes: N/A

made some changes in how the StoreScu is utilized.  In auto-routing, we now have a using() block for its usage.  In the case of processing c-moves, found a hole in how we were cleaning up.  The callback from the BeginSend() routine was a no-op, changed to clean up the scu.