Ticket: #5470 Confirmation of new A# is no longer presented during Cancel and Replace operation
Schema/Configuration Changes: N/A

Fixed the issue outlined in ticket description, but also modified the logic for checking for cancel warnings, so that it is performed prior to initiating the cancel or replace operation.  This provides a better user experience by eliminating the chance that the user will fill out the form for no reason.