Ticket: #5398 - Code Review Updates
Schema/Configuration Changes: N/A

Reworked the use of the RelativePath variable in WorkQueueUid.  This column, although left in the database for backwards compatibility, is no longer used.  We simply use the Sop Instance Uid stored in the table to calculate the path to the duplicate file.  Also did a bunch of refactoring in the code to store the some common constant strings in ServerPlatform.  These include the DICOM file extension, reconcile directory name, and duplicate file extension.