Ticket: #5398 - Code Review Updates
Schema/Configuration Changes: N/A

Integrated the failure of WorkQueueUid rows into the SopInstanceProcessor.  A FailUid method was previously used/duplicated in several locations.  Moved this into the SopInstanceProcessor.  Also reworked the return status value from the SopInstanceProcessor.ProcessFile method to not use a failure status anymore, since it wasn't really used previously.  Exceptions are expected to be thrown by the routine when an error occurs.