Ticket: #5304 - Could not change patient DOB
The mask is a bit buggy with localization and the date separator. If the date separator isn't "/" for the current localization on the server, then the mask is disabled. The field is validated against the date format for the set localization. So, even with the mask disabled, the user has to enter a valid date.