Ticket: #5186 - Patient name queries appear to take a long time Ticket: #5485 - Add DB Update Scripts for Performance milestone
Schema/Configuration Changes: Modified indexes on the Study table, made the StudyDate index be clustered and removed clustering from the Primary Key on the study table

Changed the study table to no longer be clustered on the primary key, but rather be clustered on the Study Date.  This will cause all of the rows for a given date to be together in the database.  Since most queries against the ImageServer are "today" queries, this should make the database more effective in responding to queries.

Note that on upgrade, this is basically going to cause the entire Study table to be re-written, since the order of the rows will be changed.  This will cause potentially significant delays on a large database during upgrades.