Ticket: #5839
More updates.  Cleaned up minor issue with trying to label the desktop app.  Found an issue with trying to set a config option that didn't exist.  Traced back to invalid config option in the app.config in the clearcanvas.desktop assembly.  Also added in a new splash screen that labels the product as the migration tool.  Note that this isn't the real required labeling yet, but its a start.