Ticket: #5864 - ImageServer stops if viewer attempts to retrieve studies behind firewall
Schema/Configuration Changes: N/A

Note that this problem does not happen just when the viewer is behind a firewall.  Simply having an invalid port number configured for the client on the server will replicate the problem.  Added a try/catch block around the offending exception with a log resolved the problem. Also made some adjustments to have some dicom library exceptions serializable, which was also causing issues in how the error was reported.