Ticket(s): #5289 - History not being logged
Schema/Configuration Changes: N/A

Fixed problem with the history not being logged for edits.  This was caused by the plugin attribute for the ClearCanvas.ImageServer.Core assembly not having a space in it.  The build tasks for setting the assembly version did not identify this assembly as a plugin properly because of this issue.