Ticket: #5863 - Series moves do not work
Schema/Configuration Changes: N/A

There were a number of issues with the series moves.  The WOrkQueueUid entries were not dealt with and cleaned up properly was the larges one, and a few other things had to be cleaned up.  I think there was a move from having the series be encoded in the Xml column of the WorkQueue to having them be in the WOrkQueueUid table, and testing was not done to make sure it properly worked after that change.