Ticket(s): #5859 Allow server-side custom validation rules
Pre-requisite to implementing this ticket is to create a way to eliminate multi-threaded access to the jscript engine.  The JScript.NET runtime is not thread-safe and therefore it can exhibit weird behaviour if accessed from multiple threads.  For that reason, we need a mechanism to synchronize access before the engine can be used safely in a server application which is inherently multi-threaded.

1. Added ScriptEngineOptionsAttribute which allows a script engine to specify characteristics of its usage, such as whether it must be used as a Singleton and whether access to it must be synchronized vs free-threaded.

2. Marked the Jscript engine as both Singleton and requiring synchronized access.

3. Modified ScriptEngineFactory to honour the ScriptEngineOptionsAttribute.  This required changing the public API from CreateEngine to GetEngine, since in the case of a singleton engine only one engine will ever be created.

4. Added SynchronizedScriptEngineWrapper class, which is used internally by ScriptEngineFactory as non-intrusive way to enforce synchronized access to an engine instance.