Ticket(s): #5859 Allow server-side custom validation rules
Did some refactoring so that the validation rules are now stored in the Healthcare.EntityValidationRulesSettings, instead of being stored in Enterprise.Core.Modelling.EntityValidationSettings.  It just seems to make more intuitive sense if the rules are stored as part of the model to which they apply.

This required introducing an extension point, ValidationRuleSetSourceExtensionPoint, which allows any server-side plugin to act as a source of entity validation rules.