Ticket: #6128 - Update to the 2009 version of the DICOM standard
Schema/Configuration Changes: N/A

Made a small update to the dictionary generator and introduced changes from the 2009 version of the DICOM standard.  Some additional tags and sop classes were added, etc.  Note that the 2009 version of the standard introduced Keywords for each DICOM tag.  We decided not to start using these keywords, because the keywords do not use camel casing, and because the change might cause us to have issues with auto-route rules in the imageserver that depend on the keywords we created for each tag.