Ticket(s): #6211 - Scalability and Performance enhancements
Corrected a bug introduced by the Owls merge, where the IReportingWorklistItemBroker and IProtocolWorklistItemBroker were being used by the workflow services to map a procedure step to a worklist item for the purposes of returning a worklist item in the response.  With the introduction of owls, this approach will no longer work because the owls brokers will query the owls view tables, which are not updated until after the commit.

Fix is as follows:
1. Added InitializeFromProcedureStep(...) method to WorklistItem class.  If you only want to map a single procedure step, or even a handful of procedure steps, to a worklist item, it really isn't necessary to invoke the broker to do a projection query - the worklist item can simply be initialized by accessing the properties on the procedure step and associated entities.  This is horribly inefficient because it generates many additional trips to the DB, but for one or two items it really doesn't make any difference.  Obviously this should not be used to construct a large number of worklist items.

2. Changed workflow services to call this method instead of using the brokers.

Not able to test this since I don't have a full db setup at home, but *hope* it doesn't break anything.