Ticket(s): #6363 - HL7: Reproduce existing functionality of inbound HL7 interface from UHN/Quadramed Task 6434
Reverting all extended properties collections to use Dictionary<string, string> as they were originally designed.  The idea of having an ExtendedPropertyValue type proved to provide no benefit at all due to limitations in the current HQL implementation that prevent querying on collection elements when those elements are user-defined types or components.  This can be considered a failed experiment, but fortunately one that was not too costly in terms of time invested.

Unfortunately this means more thinking around how to fulfill the needs of the HL7 requirements.