Ticket(s): #6531 The New Order, Modify Order dialogs should allow the entry by the user of a Department.
Added some validation of the performing department on the server-side.  This is done in the service layer for now... ideally it should be in the model layer, but there is a lot of validation code in the OrderEntryService that should be moved to the model layer, so it would be better to do it all at once as a separate ticket.