Ticket(s): #6211 - Scalability and Performance enhancements
Fixed a bug in OptimizedWorklistItemSearchExecutionStrategy which I just stumbled upon in the course of trying to debug something else.  The logic used to break apart the search criteria into different chunks to be executed separately was incorrect.  The code was merely checking for the existence of a PatientProfile sub-criteria, but not checking if this sub-criteria was empty or not, and this was quite possibly contributing to poor search performance.  I have now cleaned up the SearchCriteria API (previous commit) so that this type of mistake cannot be made in future.