Ticket: #6783 Investigate use of caching on folder counts
Added separate settings properties for User-dependent vs User-independent worklists.  Both are now optionally cacheable, but user-depending worklist counts are cached on the client side, where user-independent worklist counts are cached on the server.  This is because the former cannot be shared between users, while the latter can and should be.