Ticket: #6795 - About box, login screen, title bar indicate an uncertified installation
Schema/Configuration Changes: N/A

Ugh, there were actually more files that were missing.  For some reason my icons aren't changing in the explorer for TortoiseSVN, so I didn't realize it.  Here's the changes committed:

Added display code for the certification and integrated it in with the desktop app.  Also made changes to the RIS Client and ImageViewer solutions to include the clearcanvas.utilities.manifest assembly in the builds.  Introduced into the Desktop app a red label on the splash and about box if the app is not certified.  Also modified the title bar in the app to mention the same.