Technical Requirement: #6486 Implement View Maintenance Shred
In order to make the maintenance shred support having multiple concurrent instances running, the easiest way to do this was to implement the deletes using DML-style delete queries.  This commit adds support for DML-style deletes to IEntityBroker (this support was only recently added to NHibernate).

Also pulled out a non-generic base class from EntitySearchCriteria<T>, as this was needed by the shred implementation.