Ticket(s): #6925 - General clean-up of Owls code, tie up loose ends
This commit takes care of ensuring that when relationships between procedure, orders, patients, and visits change, the owls view items are updated appropriately.  Previously it was assumed that these relationships were stable over the lifetime of the entities, but with the introduction of functionality such as Merge Orders, Move Visit (via HL7), Merge Patient (via HL7), this assumption no longer holds.

1. Added an updateDirectReferences flag to a number of the view item SetXX methods, where it is possible that those methods may need to propagate the call to any direct references. The propagation rules are listed here:

ProcedureStep -> Procedure
Procedure -> Order
Order -> PatientProfile, Visit

This means for example, that if a Procedure is updated, it is possible that it was moved to a different Order, implying that it's Order reference has changed, meaning that order, patient, and visit info in the view item must be updated too.

2. Modified all XXViewUpdater classes so that the change handlers pass updateDirectReferences = true whenever this parameter exists.
3. In contrast, when updated from a tuple, this flag is passed as false, because the tuple already ensures that all components of the view item will be updated.
4. Added unit tests for WorklistViewItemBase and MwlWorklistViewItem to ensure that the "updateDirectReferences" flag exhibits correct behaviour.

5. Added PatientProfile properties to Order and Procedure classes.  This property returns the appropriate patient profile given the performing facility.