Ticket(s): #6376 Combining two orders into one
Schema/Configuration Changes: Yes. New enumeration value added to ProcedureStatusEnum.

Determined that we need a new procedure status for the so called "place-holder" procedures that are left behind when an order is merged.  The reason is that we do not want these place-holders (henceforth named 'ghost') to show up in searches, and there is currently no way to distinguish them.  Therefore, added a new procedure status GH (Ghost) and these procedures are set to this status.  This makes more sense than setting them to Cancelled, which was the previous behaviour, because they are not really cancelled.

-this allowed elimination of one TODO from owls code
-need to confirm whether this has any impact on HL7 code???