Ticket(s): #6925 - General clean-up of Owls code, tie up loose ends
Eliminated 4 TODOs related to the "most recent step" syndrome, by simply deciding to accept the risk entailed. So far we have no use cases that infringe on the boundary cases specified in the TODOs, so it seemed reasonable to choose to live with the risk (since we have no good solution).  Added other comments in place to describe the issue.