Ticket: #6977 - The server should store app information in a persistent store.
Schema/Configuration Changes: N/A

Introduced new assembly, ClearCanvas.UsageTracking.Core, which has the usage tracking code for accessing the persistent store in it.  Added a RegisterProcessor class for processing the incoming messages, and linked it in with the web service.  Not unit tested, but its a core section of code needed to process incoming messages.