Ticket(s): #7041 Number-based searches are slow, and named based searches aren't all that fast either
Modified the indexing strategy on the PatientProfile table so that instead of having separate indexes on FamilyName and GivenName, we have a single composite index on (FamilyName, GivenName).  This results in a far more efficient query plan for name-based searches involving both a first and last name.  It will make searches by first-name only incredibly slow, but the GUI should not allow these in any case, as there is no justifiable use case.