Ticket: #7247 - Integrate Monarch Code into Jin
Schema/Configuration Changes: Namespaces were changed for the Monarch specific assemblies.

Initial version of just the main monarch project ported into the Jin folder.  Renamed the namespaces to just use "MigrationTool" for namespaces and assemblies instead of "Monarch" to match how we've named things in Jin up until now.  Also included the ImageViewer solution from the Monarch Branch.  Note that this solution will probably be reworked when the actual builds start being done from the Jin folder so that the solution is a generic solution with all the project namespaces included.