Ticket: #6648 - Compression Ratios in Codecs should be based on Bits Stored rather than Bits Allocated
Schema/Configuration Changes: N/A

Updated the Jpeg 2000, Jpeg Baseline, and Jpeg Extended codecs to have their compression ratios based on Bits Stored.  Added a new property to DicomPixelData, BitsStoredFrameSize, that reports the uncompressed frame size in terms of the bits stored.  Updated the Jpeg baseline, Jpeg Extended, and Jpeg 2000 OpenJpeg codecs to use this setting when compressing frames.  Also made an update to the Pegasus codec to use this.  Finally, the license block needed to be updated for the Pegasus code.