Ticket(s): #7460 Merge practitioners/contact points must migrate orders/visits asynchronously
Added back in the logic to copy phone, address, and email information from the replaced contact point to the replacement contact point, expiring any that are not expired.  Note that this logic resides in the service layer, because this is more an aspect of the current application design (the GUI does not allow the user to select which information to keep and discard).  The MergeContactPoints method in the model layer is still extremely generic.