Ticket: #7428 - SOP Processed Compressions don't remove embedded overlays
Schema/Configuration Changes: N/A

Moved the extraction of embedded overlays out of the ImageServer code and into the DicomMessageBase.ChangeTransferSyntax method, where it really belongs.  This will cover the original defect that the SOP Processed compression rules were not removing the embedded overlays.  Also added in regression tests to test the removal of the overlays.  As part of the changes also implemented some resharper suggestions.