Ticket: #7479 - The service shall store usage information in a persistent store
Schema/Configuration Changes: N/A

Added a new field to send if the application is starting up, shutting down, or running with the usage tracking message.  Also changed the UsageTracking.Register call to include a flag if the message should be sent on the current thread or a background thread.