Ticket: #7550 - Get usage tracking server up and running on apps server
Schema/Configuration Changes: Yes

Updated the usage tracking application deployment project.  Updated the configuration of tracking to go to the https://apps.clearcanvas.ca to access the usage tracking service.  (The service is already deployed on the apps server).  Turned on the UsageTracking configuration option in the build so its enabled by default so I get some additional development usage messages going forward.  Note that this will be disabled for the "real thing" according to the current requirements.