Ticket: #7482 - The service should log “bad” phone home attempts
Schema/Configuration Changes: N/A

Changed the checked in code to connect to 4rf for usage tracking instead of apps.clearcanvas.ca.  Should work properly.  Also, made some changes in how the log4net logging is setup, made the configuration be in the web.config file instead of in a separate logging.config.