Ticket: #7178 - Introduce Component and Edition to Product Settings and Manifest
Schema/Configuration Changes: N/A

Updates for Kirk, introduced the Product/Component/Edition into the manifest and validated the values.  Note that this really isn't necessary, because the critical config with product settings is now included in the manifest and the manifest verification would fail.  This is really for readability purposes with the manifest.