Ticket: #7649 - The production service will not expose a metadata exchange endpoint
Schema/Configuration Changes: N/A

Added some final changes to the configuration to work on the production server.  Also, decided to change the virtual directory to be "Tracking" instead of 'UsageTracking' since Usage is already in the base url. This is operational on the production server at the proper url.