Ticket: #7182 - The ES user interface shall support editing the password expiration time for users
Schema/Configuration Changes: N/A

Updated the user edit screen to allow setting the Password Expiry time for a user account.  This included some updates to data contracts, etc.  Also, cleaned up the EnterpriseServer_dist.proj file as part of the update.  There were a few issues with the script not building the ClearCanvas.Executable's configuration file properly in the post build step.