Ticket: #7625 - The core framework should have a plugable license provider interface
Schema/Configuration Changes: N/A

Added new ILicenseProvider interface to retrieve licenses and supplied a default implementation that performs as the current code does, it reads and writes to the local config file.  Also changed the retrieval of the MachineIdentifier to be from the LicenseInformation class, which can now retrieve this from the ILicenseProvider when appropriate.  Needs a bit more cleanup.