Ticket: #7626 - Public Licensing Server support for retrieving the number of simultaneous users for a license
Schema/Configuration Changes:
Added new Enterprise.Core.License.LicenseSettings setting used to store the license in the config store on the enterprise server.

Added web service for retrieving license related information from the server.  Included in the changes:
 * New ClearCanvas.Enterprise.Common.License assembly that sets the datacontracts for the service
 * New ClearCanvas.Enterprise.Core.License assembly that has the implementation of the license.
 * Added the start of support for the new critical config file for the EnterpriseServer.  This is not fully functional.
 * Reworked the ClearCanvas.Common.LicenseSettings to also get a session count as part of the license retrieval.

Still needs some polish and integration into the session manager to launch the actual licensing.  Also, some thought needs to be done and further testing on authentication and the services and how the license information is retrieved, for instance, when a usage tracking message is being filled (and it requests the license key) from a shredhost.