Ticket: #7738  - Develop Kakadu based J2K Codec
Schema/Configuration Changes: N/A

Added new test tool for testing the codecs.  The test tool only compiles and works on 64-bit systems at this time.  Also made some updates to DicomCodecRegistry.  Previously, when retrieving the list of codecs, if there were duplicate codecs for the same transfer syntax, only the last plugin processed would be returned.  Now, we return all codecs (even duplicates).  Also, added a routine that allows a user to set the codec factory to use for a specific transfer syntax, so when there's duplicates, the user can override and set the codec to use.