Ticket(s): #7712 Order must have at least a visit at the ordering facility.
Schema/Configuration Changes: Yes. Server-side validation rule required

Introduction of server-side validation rule to handle this case revealed that there was a defect in the rule processing code, which would cause the "applicability rule" for every ruleset to be tested, regardless of whether the rules in that ruleset matched the rule filter.  This was an oversight in the original design and is undesirable, because there are cases where we only want to check the "low-level" rules.

The solution was to refactor the code to explicitly separate out and high level rules.  The original design had a dichotomy between "invariant" and "custom" rules, but this dichotomy wasn't particularly useful.  Therefore I removed it, and introduced instead a low-level/high-level dichotomy.