Ticket: #7746  - Add web service to retrieve enterprise server version Ticket: #7747  - Stand alone application to check EnterpriseServer version
Schema/Configuration Changes: N/A

- Added a VersionService to the Enterprise Server's shredhost and created Enterprise/VersionExecutable.

Note: VersionExecutable will be launched in the installer before any CC assemblies are deployed. Therefore, it cannot reference any ClearCanvas code. Enterprise/Common/ServiceConfiguration/Server/NetTcpConfiguration is updated to expose the metadata in debug mode so that the Version service can be consumed in this case.