Ticket: #7846 - Usage tracking store should record release type
Schema/Configuration Changes: N/A

Added Release field into usage message and storing it in the database.  Added the Release column to the ProductVersion.  This we could select to track a full release, but not a beta or development release.

Also added a db update script to help with updating the DB to include the new column.

Finally, found a small bug with the checking of lengths on the Edition field.  Cleaned this up.  (We were not checking the edition properly.)