Ticket: #7738  - Develop Kakadu based J2K Codec
Schema/Configuration Changes: N/A

Partial Commit.  Further changes to the test app to add a "test" mode to do compression/decompression and compare the output, committed the new 6.4 version of Kakadu, and made a small update to the DicomFile copy constructor to not set the ImplementationVersion tags if they're already set. Also, added a new KakaduFrameCompressor class that handles the proper disposal of the Kakadu resources with the object to make it easier to use and not have to worry about memory issues.