Ticket: #7732 - Team Edition installer will be offered for the ImageServer
Schema/Configuration Changes: Yes

Added definition for PersistentStoreVersion object, which is mapped to the DatabaseVersion_ table in the database.  Creation of the DatabaseVersion_ table will now be automatically included with the build.  Also added a shell BaseUpgradeScript that can be used when we actually introduce an upgrade script.  The Application root for updating the EnterpriseServer database will now be identical to the application root for the ImageServer.  Note also now that the PersistentStore.Version property now returns the version in the database table as opposed to the version of the PersistentStore assembly, which previously happened.