Ticket: #8119  - License Server must support username and password
Schema/Configuration Changes: N/A

NOTE:  This revision should NOT be merged into the Trunk.  The changes are already in the Trunk for Kirk and this is a temporary fix to address an issue for roll out with Euro/Fall.

Added username/password login (as set by a web.config setting) to the GUI.  Also moved the Service that the web store connects to when a purchase is made to a new URL, https://licensing.clearcanvas.ca/Generation/WebStore/Service.svc.  This was required so that authentication was not required to access the service.