Ticket: #6087 - ImageServer does not deal with private tags in c-find-rq
Schema/Configuration Changes: N/A

Updated the Query SCP to no longer fail when private attributes are included in the C-FIND-RQ.  We do not fill in these tags (as empty) in the response message at this time.  Note that updates needed to be done at each of the query levels (Patient,Study,Series,Image).