Ticket: #7315 - ApplicationLog should have nvarchar message field
Schema/Configuration Changes: Yes, modifications to ApplicationLog field.

Added database upgrade script and updated the table creation script so that the Message column in the table is now nvarchar(2000), it was previously a varchar(3000).  Also updated the Exception column to be an nvarchar field.  This should allow stack traces and error exception information to be in unicode.