Ticket: #8285  - Webviewer toolbar items not in the right order
Schema/Configuration Changes: N/A

This is interim commit

Made the changes which at least seem to fix the problem on my machine. It did not seem to work when replacing the dll in the installation with one compiled on my machine. Maybe the version didn't match, although I had modified it in the database? Will test again when the new build is made.