Ticket: #8184 - Association Abort when processing C-FIND causes service crash
Schema/Configuration Changes: N/A

Added Try/Catch statements around each of the ThreadPool calls we make to process incoming queries.  Not catching the exception was causing the crash of the service.  Also made some small changes to a test application and added a comment to NetworkBase to prevent future issues, potentially.