Ticket: #7521 - Multiframe embedded overlay extraction could cause exception
Schema/Configuration Changes: N/A

Simple fix for multi-frame overlay.  Expanded the length of the overlay array to the proper size.  Since this was such a simple fix for the problem, decided to go with this for now instead of integrating the ImageViewer code with the DICOM toolkit at this time.