[Requirement: #8747] - Expansion of Authority Groups to support fine grained access [Requirement: #8748] - Add authority group changes for desktop GUI
Initial changes in Xian for fine grained access support.  Updated the desktop
gui for authority groups to allow adding a description of the authority group
and setting the authority group for restricting data access.

Also, had to slightly rework the data contracts for Authority Groups so there
is a read service that doesn't require authentication for performing the
sync service that will eventually be developed to catch when authority groups
are deleted.