[Bug: #8514] - Inconsistent internationalization of Server Rules
Updates to internationalize tooltips.  These changes were needed in McCoy to
support McCoy being internationalized, so did the changes in the default branch
and will be merging to McCoy.

While I was at it, searched through all the Tooltips in the ImageServer GUI and
found tooltips on the Move Study and Move Series pages that were not
internationalized.  Also found a place where a WorkQueue page related tooltip
was not internationalized.

Note that this does not directly fix the referenced ticket, although its related to it.