[Requirement: #8754] - ImageServer Web GUI Authority Group & Users Support for new fields
Changed the labeling of the User Groups page to Authority Groups to match the
Desktop GUI.  Also updated the UserGroup page to include editing of the
new description and DataGroup fields and made some general improvements to this
page.

Note, this commit will require new screen captures for the ImageServer manual.