[Requirement: #8536] - Protocol matching
Tweaked the matching algorithm so that a score of 0 is considered a match.  This provides a simple means of handling the case where the user clears all the match criteria: this will now result in a global protocol ie. one that matches everything, but always appears at the bottom of the list.