[Requirement: #8549] - Hanging Protocols will be stored in a local cache
Changed the definition of IHpContributor so that the ContributorId property is a Guid rather than a string.  The purpose of this is just to make it more robust, since before it wasn't really clear what format the GUID string should be in.

Note that this change invalidates all previously serialized protocols. Therefore developers should delete their hp.sdf file and any protocols in their enterprise server.