[Requirement #8759] - WebPortal Search for Studies
Did a bit of refactoring to move the Controls folder into the Views folder,
since the ViewModels for the controls were already stored in the ViewModel
folder.  This matches up better.

Also, implemented double click of a row on the Search to navigate to the new
Images view and populate the ViewModel for the Images page.

Finally, fixed the old bug I was truggling with on the Search page with
having to reset the list of found items on the Search page.  The search now
appears smoother.