[Requirement #8758] - Home Page and Most Recent Patients View
Updates to get the Announcements section on the home page working.  Made
updates to the build process for importing of configuration settings to work
through the ClearCanvas.Executable.  Had to update the Executable to be
.NET 4 based because some of the settings to be imported were in a .NET 4
assembly related to the portal.  Note, we may have to add a new Executable
in the future specifically for .NET 4 if we don't upgrade the other apps.  Also
did some general cleanup.