[Requirement #8761] - Launch Webstation from Search Screen or Home Page
General code cleanup, did some changes so some events related to the buttons
are released properly and made some additional resharper cleanups.  Also
fixed an issue so the Images page is cached, so when you navigate back to it
the web viewer is still displayed.  Opening a new instance of the web viewer
still doesn't work yet.