[Requirement #8761] - Launch Webstation from Search Screen or Home Page
Made some changes to have popups display properly on the Webstation.  Note that
the context menu still is not working properly, for some reason the
styles are not being applied properly for the ContextMenu.  Note sure what the
cause is yet.  Also made some general code cleanup as part of the
investigation.