[Requirement #9035] - Add method for copying existing rule
Introduced the Copy button to the DataRules configuration page.  This will
now open the AddEdit dialog in a copy mode and with the ability to save
the rule.

Also, during testing noticed that a prior merge must have caused a localized
string for the ServerRules page to be missing.  Fixed this issue also.