[Requirement #8772] - Authority Group adjustments for WebPortal
Fixed the ClearCanvas.Executable configuration so that the authority group
plugins are disabled from the RIS.  Also added a new Login token for the
WebPortal to allow a user to login the portal.  Also cleaned up the authority
group import so the Description and DataGroup fields are imported properly
if they're set with the authority group descriptions.  Made some other general
cleanup as part of this fix.