[Task #9033] - Unit tests [Requirement #8999] - SUV calculation shall support different device manufacturers
- Fixed a serious bug in PhilipsSUVParameters when extracting the SUV scale factor value from private tag. Problem was found when testing with real Philips images although all unit tests passed.
The problem arose because private tags are handled differently by the dicom library depending on how it is used.

In unit tests, images are created on the fly and private tags with correct VR are added to the collection. Their values when extracted are always correct.
However, when the image is loaded from disk, DicomStreamReader uses DicomAttributeUN for private tag. We need to parse the raw value ourselves because DicomAttributeUN does not support any GetXXXX() methods.

- Also refactor some of the code so that dicom values are loaded on demand and moved some Philips specific code from StandardSUVParameters into PhilipsSUVParameters.
- Updated the unit tests

Schema/Configuration Change:  None