[Requirement #9198] - Eliminate delays between login and usable application
The initial call to the phone home service was being done synchronously, contributing to the overall start up delay.  Changed the code so that the initial call is done on the timer callback, 10 seconds after the application starts.  This way, it does not interfere with the GUI thread at all.