[Requirement #9198] - Eliminate delays between login and usable application
Removed the whole "secondary app domain" thing from the plugin loading process, since it really wasn't necessary.  Its purpose was to conserve memory by unloading any non-plugin dlls, but realistically any non-plugin dlls found in the plugins directory are going to be loaded anyways because they are generally referenced by a plugin.

In the process of making this change, I essentially rewrote the entire PluginManager class, because it was needlessly complicated.