[Requirement #9199] - Ensure window drawing occurs prior to beginning intensive initialization activities
Modified the StudyBrowserComponent so that searches are performed on a background thread rather than on the GUI thread.  This way, the initial query to populate the list with the local studies does not block the GUI from drawing.

While the search is executing, the StudyBrowserComponent and the ServerTreeComponent are disabled, and the SearchPanelComponent switches to a mode where most of its controls are disabled but the Search button changes to say Cancel, allowing the user to cancel the operation if desired.

Note that cancelling does not actually cancel the DICOM query, but simply abandons waiting for the results.