[Requirement #8759] - Automatic Timeout of the GUI
This includes a bunch of refactoring of how events are handled:
 * Completed the base refactoring of the old ServerEventDispatcher.  The code is
now split between a ServerEventMediator class and a ServerMessageSender class.
 * The new ServerMessageSender basically wraps the WCF client for sending.  The
wrapping allows for some well definied error handling and fault handling.
 * The ServerMessagePoller has been adapted to use the ServerMessageSender class
when it polls, instead of the direct access to the WCF client.
 * The ServerEventMediator has the dispatcher code from before. Also,
refactoring has been done such that error messages are sent through events
to the dispatcher and to the WebPortal or Webstation client classes.
 * The displaying of failure dialogs has been refactored out of the ImageViewer
class so that it can be done differently in the WebPortal and Webstation.
 * Minimized the use of the ApplicationContext.Current class and changed the
core webstation controls to keep a reference to ServerEventMediator.  The
mediator class is now self contained in the ImageViewer, and is not a static
resource, so we could open multiple webstations within the same silverlight
app if need be.

There's still some more polish that needs to be done and testing with the
ImageServer's Webstation, but it works in the WebPortal.