[Requirement #8759] - Automatic Timeout of the GUI
Additional basic changes to get the timeout to work, which is now operational.
Introduced the EventToCommand class from the MVVM Light toolkit.  This class
helps bind events to the ViewModel, so the event handler doesn't have to go
into the code behind.  Also added in WebPortalViewModel, which handles the
timeout of the application.  Finally, did some refactoring work and consolidation
of the DelayedEventPublisher, this was located in two places and has been
consolidated.

Overall, its working, but could use a bit of polish, and checking to make sure
the server side application is cleaned up properly.