[Requirement #8759] - Automatic Timeout of the GUI
Cleaned up how a connection is shutdown and application closed on the server.
I was having some problems with late messages arriving after I'd quickly close
the viewer and opened a new one in the WebPortal.  Now, once we issue a stop
application request, we basically ignore any communication to or from the server.
This eliminated some issues with opening/closing.  Also, fixed some other
small issues that were generated errors in the log related to timeouts.  Removed
some debug statements, and finally added in some better client side logging to
track what is occurring on the client.  This helped debug the above issues.