[Requirement #9229] - PDFs viewable within WebPortal
This update includes a number of changes for viewing PDFs:

- Introduced a mechanism to display PDFs on a new Reports page.  Right now a
generic PDF from the web is being displayed, but the framework is there for
displaying our real pdfs.
- Introduced new WADO PDF mime type handler.  The ImageServer code will now
return the raw PDF document from an encapsulated PDF when the mime type is
asked for.
- Added a query mechanism in the ria services to get if a study has
encapsulated pdfs associated with it.
- Added new PDF Http Handler. This will be used to translate any request for a
PDF to the web server into a wado request to the imageserver.
- Introduced some client WADO code into the web server.
- Updated the StreamingClient class in the dicom library to have a method for
requesting a PDF.