[Requirement #9229] - PDFs viewable within WebPortal
Cleaned up a few issues with the PDF viewer.  The PDF now is not visible when
a pop-up message is displayed by the web portal.  It comes back when the
popup is dismissed.  Also, fixed an issue where the list of reports was not
being dismissed.  Also, finally removed the pop-up that was being displayed
when the system was testing the connectivity speeds for the web viewer.