[Requirement #8645] - Perform Design Reviews
A bit of research shows that the way I was propogating property notification
using some MVVM Light toolkit code would not work properly when the silverlight
app was compiled in release mode.  Updated all the properly update methods
to use an alternate method for determine the property name to propagate.