[Requirement #8765] - Develop About Page
Fixed bug introduced with prior commit of this file.  Added a 'v' in front of
the version, which ended up impacting Usage Tracking, where the version sent
to the server had the v.  Changed so that only when the Name & Version is
returned together is the v added in front of the version to match out code.