[Bug #9276 ] - Pressing ENTER in a WebPortal search field does not apply the latest query parameters
This was caused by a binding problem, where the KeyUp event was occurring before
the TextBox's Text field was bound to the ViewModel for the search page.  So,
the search was performed before we had the new value to the text field.

Added a new class, TextBoxEnterButtonInvoke, that allows binding the enter
key being processed in a TextBox to a button being pressed.  Also, added code
to force the TextBox's Text field to be re-bound when the enter is pressed
before the button is invoked.

Discovered that the WOrkstation is filling in SeriesDescription instead of
DocumentTitle.  Made a small adjustment in the code to use series description
instead of document title when we have multiple reports for a study if the
document title isn't filled in.