[Defect #9315] - Administrative Services encounters exception and failed to install
Excluding the class from obfuscation does not seem to fix the issue.  Seems like the obfuscator simply cannot handle string arrays in attributes, even if the entity to which the attribute is attached is excluded from obfuscation.

To work around this issue, changed the Formerly property to a string instead of a string array.  Multiple values can still be assigned, by combining them all into one semicolon separated string.  Hopefully this does the trick.