[Defect #9312] - Default hanging protocol export save path can cause the manifest to break
Added some code into the Desktop framework to check and prevent users from choosing to save something to the install directory via a file save dialog, since this is generally considered a bad practice.

This is the default behaviour.  Developers can override this behaviour on a case by case basis by setting the PreventSaveToInstallPath flag to false in FileDialogCreationArgs.