[Defect #9254] - Opening images in webstation presents only a black screen until images are loaded
Added a new busy indicator while the content is loaded.  This is displayed
until the app starts and is show.  We now have a shared busy indicator that
can be used in any of the silverlight assemblies or projects.

Added the System.Web.Extensions.Design assembly into the web folder.  Hopefully
this should remove the need to have the ajax extensions installed.