[Defect #9317] - ImageServer installer failed to upgrade over the previous Team Edition
Did some more thorough testing, with a test app against a real installed
version and I think I have the upgrade finally working.  There was an issue
with the script itself not being an embedded resource and with the format of the
script also not being supported by how the upgrade script was being rul for the
enterprise version.