[Defect #9412] - Possible Memory Leak in WebPortal
A bunch of changes in an attempt to eliminate the memory leak.  Found a number
of issues with the ServerEventMediator and how it was being cleaned up, we
were leaving background threads running.  The issue is still not solved,
but the code should be a bit cleaner.  Also cleaned up the Dispose pattern
on a number of classes.