[Defect #9416] - Editing study in ImageServer causes person name suffix to disappear
The PersonName class did not have a Suffix available.  Added the Suffix.  It
was already supported by the ComponentGroup, the property was just not
available.

Also, updated the GUI itself in the PersonNameInputPanel, the Suffix was not
being bound to the proper control, probably due to it not being available
through PersonName.