[Requirement #9559] - Study Details Support for Vet Tags
Finally cleaned up and have the Add Data Access working.  Made an update to the
code generator so that we now make all XmlDocument fields have an internal
member, and we call add a NonSerializable attribute to the internal member.
This allows the Study object to be serializable, because the XmlDocument is
now ignored and not serialized.

Did some additional refactoring with the StudySummary objects.