[Requirement #9559] - Study Details Support for Vet Tags
Added ResponsiblePersonRole to the study details page, as per the TR.  Also,
made a bunch of cleanup to the extensions for populating inserts and
queries to be a bit cleaner.  Note that ResponsiblePersonRole can now be
queried as a return key via dicom with this change.  Will update the
appopropriate TR.