[Requirement #9564] - Limit places where access is granted to only show the Data Access authority Groups that have access to a partition
Updated all of the dialogs for granting access to take into account if an
authority group has been granted acces to a partition or if it has access to
all studies on a partition when displaying it in the list of authority
groups to grant access.  If the Authority group has access to "All Studies",
it is not listed in the list due to it already having access to everything.
If it has access to All Partitions, it is automatically included, or if it
is granted access to the specific partition it is put in the list.