[Requirement #9536] - Ensure Build Script repository is managed properly
Attempt at getting the Manifest to work properly for the Vet edition.  Pulled
out the themes directory for the ImageServer into a dedicated file and updated
the community & Team solutions to use this.  Introduced a new Vet theme
manifest input file, and copied the Default theme and adjusted for the Vet
theme.

Also found a small problem with a Team edition license related assembly being
in the community build, and adjusted for this appropriately.